Update BlackGuard.txt
Paeon the Healer with his herbs of the River Nile, picked up the art of making panaceas from Red Dragon Isle. Powder extracted from compounds  synthesized to coat the great pyramid with bubbly tooth enamel, so that this could be imprinted with hieroglyphics was remembered by the Arabians for a very long time.